Auton uses threads now - not sure if that makes it easier to code or not

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Templates/Template_V4/DoNotChange/Robot.java
@@ -143,7 +143,7 @@ public class Robot extends Robot_Logic {
     }
 
     public void turnWithEncoders(double degrees) {
-        turnEncoders((int) (degrees * inchPer90 / 90) );
+        turnEncoders(degrees * inchPer90 / 90.0);
     }
 
     public boolean MotorsBusy() {
